Created a gem.  Have added instructions for same.  Downloaded redsafe gem for testing.  Saw there had been some downloads before I could begin testing.  Sorry.  I'm learning.  Hope you profit from that.  My current problem is the dynamic constant.  Will send update later.
Made some updates.  Details are below.

1) Use of const_set
Setting variables equal to constants did not work for me.  I got a dynamic constant error.  Found that: "const_set" was required to set constants.

a) const_set(TIME_FMT, '%Y-%m-%dT%H:%M:%SZ')
b) const_set(TIME_STR_LEN, 20 )
c) const_set(TIME_RANGE, 60*60*5)

2) +options[:remote_name]+ is deprecated, use :via instead

Finding documentation for datamapper is not satisfactory.  Some is available, but there is not enough to increase knowledge of the subject.  Found the link below was helpful with flocating changes in datamapper.  The error message stated what to do, but more information is preferable.

http://sick.snusnu.info/2009/06/03/migrating-to-datamapper-0100/

3) http://Gemcutter.org Yank Policy

Didn't realize that once a gem is uploaded to Gemcutter, that version cannot be changed nor can it be deleted.  The reason for this is that others might be using your gem for their purposes, and they would be placed in a bind if the gem were removed.  Think it would e helpful if a grace period were allowed for developer to remove gems at will.  The current version of redsafe is 0.1.6.  As you can see quite a bit of testing has transpired.  The version should really be:  0.0.1.  Not enough progress was made to justify the high number, but rules are rules.